+++ src/main/java/dev/andrylat/carsharing/dao/mappers/CarModelMapper.java
@@ -12,10 +12,10 @@ public class CarModelMapper implements RowMapper<CarModel> {
     public CarModel mapRow(ResultSet resultSet, int i) throws SQLException {
         CarModel carModel = new CarModel();
 
-        carModel.setId(resultSet.getInt("id"));
-        carModel.setBodyId(resultSet.getInt("body_id"));
-        carModel.setBrandId(resultSet.getInt("brand_id"));
-        carModel.setFuelId(resultSet.getInt("fuel_id"));
+        carModel.setId(resultSet.getLong("id"));
+        carModel.setBodyId(resultSet.getLong("body_id"));
+        carModel.setBrandId(resultSet.getLong("brand_id"));
+        carModel.setFuelId(resultSet.getLong("fuel_id"));
         carModel.setName(resultSet.getString("model_name"));
         carModel.setEngineDisplacement(resultSet.getDouble("engine_displacement"));
         carModel.setGearboxType(resultSet.getString("gearbox_type"));
